Change app launcher package name
From io.vertx.core.application to io.vertx.launcher.application
