+++ application/src/main/java/io/vertx/launcher/application/VertxApplicationHooks.java
@@ -9,7 +9,7 @@
  * SPDX-License-Identifier: EPL-2.0 OR Apache-2.0
  */
 
-package io.vertx.core.application;
+package io.vertx.launcher.application;
 
 import io.vertx.core.Verticle;
