+++ application/src/main/resources/META-INF/MANIFEST.MF
@@ -1,1 +1,1 @@
-Automatic-Module-Name: io.vertx.core.application
+Automatic-Module-Name: io.vertx.launcher.application
