+++ application/src/test/java/io/vertx/launcher/application/HttpTestVerticle.java
@@ -9,7 +9,7 @@
  * SPDX-License-Identifier: EPL-2.0 OR Apache-2.0
  */
 
-package io.vertx.application;
+package io.vertx.launcher.application;
 
 import io.vertx.core.AbstractVerticle;
 import io.vertx.core.json.JsonObject;
