+++ application/src/test/java/io/vertx/launcher/application/TestHooks.java
@@ -9,11 +9,9 @@
  * SPDX-License-Identifier: EPL-2.0 OR Apache-2.0
  */
 
-package io.vertx.application;
+package io.vertx.launcher.application;
 
 import io.vertx.core.Vertx;
-import io.vertx.core.application.HookContext;
-import io.vertx.core.application.VertxApplicationHooks;
 
 public class TestHooks implements VertxApplicationHooks {
