+++ application/src/test/resources/META-INF/MANIFEST-Missing-Main-Verticle.MF
@@ -1,2 +1,2 @@
-Main-Class: io.vertx.application.VertxApplicationTest$TestVertxApplication
+Main-Class: io.vertx.launcher.application.VertxApplicationTest$TestVertxApplication
 Main-Verticle: org.acme.MissingVerticle
