+++ application/src/test/resources/META-INF/MANIFEST-No-Main-Verticle.MF
@@ -1,1 +1,1 @@
-Main-Class: io.vertx.core.application.VertxApplication
+Main-Class: io.vertx.launcher.application.VertxApplication
